phase/4 update BlogTag list
